+++ src/fbchat_v2/_core/__init__.py
@@ -22,6 +22,6 @@ from importlib.metadata import version, PackageNotFoundError
 try:
     __version__ = version("fbchat-v2")
 except PackageNotFoundError:
-    __version__ = "2.3.0"
+    __version__ = "2.3.1"
 
 __all__ = ["_session", "_utils", "_facebookLogin", "__version__"]