+++ src/fbchat_v2/_features/_facebook/__init__.py
@@ -28,4 +28,7 @@ __all__ = [
     "_registerOnProfile",
     "_search",
     "_unFriend",
+    "_archivePost",
+    "_deletePost",
+    "_reactionPost",
 ]